+++ src/intermission.rs
@@ -1,3 +1,4 @@
+mod card_view;
 mod character_scene;
 mod equipment_view;
 mod reward_scene;
